Sends SMS when order placed
When order is placed, the waitingOrder object is stored in memory.
The restaurant is notified and given instructions to set the time to prepare the order.Sends SMS when order placed

When order is placed, the waitingOrder object is stored in memory.
The restaurant is notified and given instructions to set the time to prepare the order.